training/pyproject.toml
[project]
name = "trainer"
version = "0.1.0"
description = "LLM finetuning pipeline for Go framework documentation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core ML
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "peft>=0.7.0",
    "trl>=0.7.0",
    "bitsandbytes>=0.41.0",

    # Data
    "datasets>=2.15.0",
    "safetensors>=0.4.0",

    # Utilities
    "pyyaml>=6.0",
    "requests>=2.31.0",

    # Experiment tracking
    "wandb>=0.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
trainer-synthesize = "trainer.cli:synthesize"
trainer-train = "trainer.cli:train"
trainer-evaluate = "trainer.cli:evaluate"
trainer-export = "trainer.cli:export"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trainer"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
